+++ yaskoam.mrz2.lab2/src/main/java/yaskoam/mrz2/lab2/functions/PeriodicSequence.java
@@ -6,19 +6,19 @@ import java.util.Map;
 /**
  * @author Q-YAA
  */
-public class PeriodicSequence extends Sequence<Long> {
+public class PeriodicSequence extends Sequence<Integer> {
 
     private static final int PERIOD_LENGTH = 4;
 
-    private static final Map<Integer, Long> NUMBER_AT_PERIOD_POSITION = new HashMap<Integer, Long>() {{
-        put(0, 1l);
-        put(1, 0l);
-        put(2, -1l);
-        put(3, 0l);
+    private static final Map<Integer, Integer> NUMBER_AT_PERIOD_POSITION = new HashMap<Integer, Integer>() {{
+        put(0, 1);
+        put(1, 0);
+        put(2, -1);
+        put(3, 0);
     }};
 
     @Override
-    public Long apply(int x) {
+    public Integer apply(int x) {
         int positionInPeriod = x - (x / PERIOD_LENGTH) * PERIOD_LENGTH;
         return NUMBER_AT_PERIOD_POSITION.get(positionInPeriod);
     }
